removed empty helpers and moved all view helpers to apphelper
