--- app/helpers/account_helper.rb
@@ -1,2 +1,0 @@
-module AccountHelper
-end
