--- app/helpers/case_studies_helper.rb
@@ -1,2 +1,0 @@
-module CaseStudiesHelper
-end
