--- app/helpers/citations_helper.rb
@@ -1,2 +1,0 @@
-module CitationsHelper
-end
