--- app/helpers/collection_helper.rb
@@ -1,2 +1,0 @@
-module CollectionHelper
-end
