--- app/helpers/companies_helper.rb
@@ -1,2 +1,0 @@
-module CompaniesHelper
-end
