--- app/helpers/contacts_helper.rb
@@ -1,2 +1,0 @@
-module ContactsHelper
-end
