--- app/helpers/design_methods_helper.rb
@@ -1,5 +1,0 @@
-module DesignMethodsHelper
-	def glyphyify(value)
-	   content_tag(:span, "", :class => "glyphicon glyphicon-#{value}")
-	end
-end
