--- app/helpers/discussions_helper.rb
@@ -1,3 +1,0 @@
-module DiscussionsHelper
-
-end
