--- app/helpers/main_pages_helper.rb
@@ -1,2 +1,0 @@
-module MainPagesHelper
-end
