--- app/helpers/method_case_study_helper.rb
@@ -1,2 +1,0 @@
-module MethodCaseStudyHelper
-end
