--- app/helpers/method_categories_helper.rb
@@ -1,2 +1,0 @@
-module MethodCategoriesHelper
-end
