--- app/helpers/set_helper.rb
@@ -1,2 +1,0 @@
-module SetHelper
-end
