--- app/helpers/user_sessions_helper.rb
@@ -1,2 +1,0 @@
-module UserSessionsHelper
-end
