--- app/helpers/users_helper.rb
@@ -1,2 +1,0 @@
-module UsersHelper
-end
